Image_Scrapper
Image_Scrapper is a Django web application written in Python, in which
you would enter the URL for a website and then in the back end it would
crawl that website, and display all the images on that site on the front
end. The user can also specify how deep the links can crawl.